Examples: 修正数据库示例，增加Int64的支持。

+++ trunk/Yuki/Examples/Database/MailCSharp/Src/PostgreSql/PostgreSqlDataAccessManual.cs
@@ -14,7 +14,7 @@ namespace Database.PostgreSql
 {
     public partial class PostgreSqlDataAccess : IDataAccess
     {
-        public List<String> FromMailAttachmentSelectManyForNameById(Int Id)
+        public List<String> FromMailAttachmentSelectManyForNameById(Int64 Id)
         {
             var cmd = CreateTextCommand();
             cmd.CommandText = @"SELECT ""name"" FROM ""mailattachments"" WHERE ""id"" = @id";
